fix(exp): 经验值 long 化防溢出
- Player.exp int→long（DB 累计经验可超 21 亿，intValue() 截断导致负数落库）
- PlayerService 读取不再 .intValue() 截断
- CombatService 击杀经验累加 long
- MonsterAOI.onMonsterDeath 签名 long（S2C 报文仍 int32，击杀经验安全强转）

+++ pt-game-server/src/main/java/org/jpstale/server/game/model/Player.java
@@ -20,7 +20,7 @@ public class Player {
     private String name;
     private int job;            // 职业
     private int level;
-    private int exp;
+    private long exp;           // 累计经验（对齐 getExpForLevel 表，long 防溢出）
     private int gold;
 
     // 属性
